Revert "refactor: entrypoints (#245)"
This reverts commit 67718f0376271067a786dd56a8925e81f2e86460.
