--- packages/unity/types.ts
@@ -1,8 +1,0 @@
-export type UnityLoaderType = {
-  instantiate(divId: string, manifest: string): UnityGame
-}
-
-export type UnityGame = {
-  SendMessage(object: string, method: string, ...args: (number | string)[]): void
-  SetFullscreen(): void
-}
